08_Deployment/Frontend/TreeTrace/requirements.txt
torch
torchvision
pytorch-lightning
opencv-python
numpy
pandas
scikit-learn
matplotlib
seaborn
plotly
streamlit
fastapi
uvicorn
pydantic
pillow
tqdmp
pyyaml
jupyter
notebook
